Use tarball instead of git repository
Apparently, the implementation of how it fetches the source code is different in CMake. For tarballs it assumes that the tarball never changes, and so checks if the tarball was already downloaded previously before downloading it again. For Git repositories it always clones again. According to awhiemstra.

Contributes to issue CURA-2497.