Revert "Patch ComTypes with a small fix"
This reverts commit 20f8051a7cc9b21f07dac4973ab7639029136403.
It didn't work on Windows since there is no 'patch' on Windows. To fix this we'd have to pull in GNU's patch application as an additional dependency and that takes this to an additional level, especially seeing as to how I'm doing all this from an Ubuntu machine...

Contributes to issue CURA-3823.