Add build python windows BAT
Need to pass the argument "/p:PlatformToolset=v140" which needs to be
strictly put in double quotes. Otherwise the "=" will cause problem in
Windows batch and it will be evaluated as "/p:PlatformToolset v140".
Pass it through CMake via a command always seem to result in
"/:PlatformToolset v140", so I put this command in a dedicated bat
script.