Add patch to fix installation directories on Windows
On Windows the installation directories were incomplete as it didn't find a library directory. This adds 'include(GNUInstallDirs)' which hopefully resolves the problem.
This is done as a complete replacement of the CMakeLists.txt file since Windows has no patch executable that applies unified diff files.

Contributes to issue CURA-7259.