Add patch for finding Clipper for libnest2d on Windows
On Windows the environment variable doesn't get set, and as a result Clipper doesn't get found. Instead we'll patch the find script in libnest2d. Quite hacky, but effective. We add that it searches also in CMAKE_INSTALL_PREFIX for the library and the headers. Other solutions (changing CMAKE_PREFIX_PATH via a -D argument, actually setting the variable with setx, etc.) were ineffective. It's weird.

Contributes to issue CURA-7259.