Add importlib-metadata dependency
This is a package that used to be a real dependency of cx_Freeze, causing it to be installed automatically. It no longer is, since it's included in CPython since Python 3.8. However cx_Freeze doesn't seem to have updated the way in which they import this library, since it then got renamed from importlib_metadata to importlib.metadata. The cx_Freeze source code (even the latest version) still imports importlib_metadata.
This package, which used to be the real package, now provides backwards compatibility for projects like cx_Freeze.

Contributes to issue CURA-7900.