On macOS Catalina install scipy via pip.
Scipy is compiled using cmath files and functions from the system. However, the build fails because it cannot find certain functions in the global namespace in those Xcode files.
This seems to be a problem with Xcode, because both the `/Library/Developer/CommandLineTools/usr/include` and `/usr/local/include` are in our CPATH. The solution is to remove the latter, but it is strictly required for Python to be built. Hence, the easiest and also fastest solution is to install scipy via pip on macOS.
See the error here: https://stackoverflow.com/questions/58628377/catalina-c-using-cmath-headers-yield-error-no-member-named-signbit-in-th