Use requirements.txt to install python deps
Using a requirements files allows us to install all packages with a
single command. Specifying the hashes for each dependency.

By using environment-markers in the requirements file exceptions can
be made for each platform.

On Linux we used to build Scipy from scratch, this was done because
older versions didn't have a pre-build wheel and compiling it via
pip gave problems with the gfortran compiler. The version of pip
which we use now, prefers wheels above source and the Scipy version
has a pre-build wheel available.

I'm still a bit confused why we need custom wheels for Scipy and
shapely on Windows and why the patch version for each Shapely
various per OS.

Contributes to CURA-8403