Add patch for libffi version used by Python
In the Python 3.8 source code they download libffi binaries from their Github repository at github.com/python/cpython-bin-deps. They recently updated cpython-bin-deps so that it no longer contains libffi-7.dll and libffi-7.lib, but now contains libffi-8.dll and libffi-8.lib. The source code for Python 3.8.10 still refers to libffi-7.lib and .dll though, because we didn't update CPython. And even if we did update CPython (to a version they didn't release yet) it would still not be fixed because they forgot to update this file. Bit of a broken situation there.
Let's try this first; it would include security fixes in libffi. If this doesn't work we'll have to affix the commit ID it's checking out in the build script in Python's source code.