Force Python to download and install OpenSSL v1.1.1l
By default, the python3.8 build configuration still downloads and installs OpenSSL v1.1.1k from the cpython-bin-deps repository. As a result, we will have to force it to download OpenSSL v1.1.1ll, which means that we need to change the expected version in the get_externals.bat and the python.props files.