Add macOS-specific post build commands
On macOS the loader instructions are using the relative `@rpath` location for linking and loading the dependencies of `geos` and `pynest2d`. As a consequence, copying them to the eventual `Ultimaker Cura.app` bundle will fail. Thus, we need to specify the absolute paths for those dependencies using macOS's internal `install_name_tool` command.