Remove the DYLD_LIBRARY_PATH
On macOS PyQt fails to build because it perceives the Qt installation as incorrect or broken. After investigating this happens because we set the `DYLD_LIBRARY_PATH` variable before configuring it.
It used to work just fine, but now even our build jobs from Rundeck are failing because of this, so most probably something must have changed in our system (perhaps latest upgrades are messing things up?).