Provide install prefix as prefix for packages
If packages have to be compiled, they have to be able to find their compilation dependencies.
In particular, we had to compile Shapely, and Shapely didn't find its compilation dependency Geos. Geos was installed in the install prefix, but easy_installation was looking for it in the system default folders on MacOS (/opt/homebrew, /usr/local, etc). Hopefully this fixes that. It still works on Linux, but on MacOS I can only test on the build server.

Contributes to issue CURA-8592.