Update cx-Freeze to 6.9
This is currently the most recent version.
We're updating this because the previous version is giving an error when packaging on Linux for Python 3.10. I couldn't find the error online, but I found something similar where the issue was that cx-Freeze didn't support Python 3.6 yet. Also, the official documentation says that 6.9 is the first cx-Freeze version to support Python 3.10.

Contributes to issue CURA-8592.