Update Pip and Setuptools
Hopefully this resolves the build issues we're getting on Windows because of Distutils missing a symbol. I've read that this would need Pip to be at least 19.2, but we're already above that. However we might need to update Setuptools too since that is where the issue was actually occurring.

Contributes to issue CURA-8592.