Add cx-Logging as requirement of cx-Freeze
On Windows (only), cx-Freeze requires this dependency. It requires cx-Logging>=3.0. However since we require hash checking of our dependencies, this breaks. It says it needs a specific pinned version (== instead of >=). And it'll need a hash for that version. So we'll specify the version to use, here.
I'll add it for all platforms, not just for Windows where it's a requirement, so that our build system is more consistent across platforms.

Contributes to issue CURA-8592.