Add new binaries for Scipy instead of Gohlke's
Gohlke's Scipy build seems to give us a circular import on Windows. The cause of this is unknown, but it's occurring simply when we import scipy.spatial. It could be that this is caused by a similar problem to https://github.com/scipy/scipy/issues/4048 .
We'll have to use the non-MKL builds for now. This means that Scipy operations (like getting the convex hull) could become slower on Intel CPUs on Windows, which is quite a common configuration. I don't see a better solution to this though.

Luckily, there are now binaries for CPython 3.10. I don't know why I didn't see them before since it says they were uploaded towards the end of November? Maybe I overlooked them. Or maybe I was using Hashin and used that wrongly. Or maybe they weren't there yet somehow.

Contributes to issue CURA-8592.