Remove links to current environment
After switching to CentOS 8, libSavitar could not be configured anymore due to a linker error.
It happened because we are linking to the currently installed version of python and several system lib folders, which are messing things up. Since we're building on docker now anyway no such linking needs to take place anymore. Same error was occurring on Arcus and pynest2d.