Delete certain Qt plug-ins after downloading them
These cause the builds to fail if they are present.
Let's hope we don't need them. This hasn't yet been properly tested at this point since the build segfaults on something else.
The NativeStyle hasn't been tested yet whether it truly fails the build. It seems to have the same problem though, of having a binding there but not in the Qt implementation.

Contributes to issue CURA-8640.